asdfgfdd

+++ src/model/Receitas.java
@@ -2,30 +2,30 @@ package model;
 
 public class Receitas {
 	
-	private float donativos,
+	private double donativos,
 				  quotas;
 	//private String data;
 	//private int data;
 
-	public Receitas(float donativos, float quotas) {
+	public Receitas(double donativos, double quotas) {
 		super();
 		this.donativos = donativos;
 		this.quotas = quotas;
 	}
 
-	public float getDonativos() {
+	public double getDonativos() {
 		return donativos;
 	}
 
-	public void setDonativos(float donativos) {
+	public void setDonativos(double donativos) {
 		this.donativos = donativos;
 	}
 
-	public float getQuotas() {
+	public double getQuotas() {
 		return quotas;
 	}
 
-	public void setQuotas(float quotas) {
+	public void setQuotas(double quotas) {
 		this.quotas = quotas;
 	}
